Expose which variables to consider for system FCT to user
This commit exposes the variables that are considered for system FCT to
the user via an input file construct. Example:

sysfctvar 1 2 3 4 end

The above ignores the density when computing a limit coefficient for
FCT. The default for system FCT (if enabled) is to consider all of
density, momentum, and total energy.

The system character of FCT can be enabled by 'sysfct true', which is
now true by default for single-material compressible flow equation
systems. Also, the system character for FCT is now configured within a
compflow ... end block, which allows assigning and configuring a system
character to a specific compflow system.

In summary, the FCT-related defaults are:

fct true
ctau 1.0

compflow
  sysfct true
  sysfctvar 0 1 2 3 4 end
end

This commit also extends FluxCorrector to account correctly for multiple
compflow systems both for what variables are considered as a system
(there can be multiple) as well as velocity for setting symmetry BCs.
This is tested using two compflow ... end blocks, which now breaks later
and not due to setting up FluxCorrector.

Since the default for sysfct (affecting all compflow/diagcg) tests have
been changed from false to true, regression tests have als been updated.
One non-trivial update is compflow_euler_sodshocktube_u0.8_migr, which
now does not do exodiff, only diag diff. It was too hard and error-prone
to set up a test that works with SMP, migration, virtualization, and
reproducable.