Expose FCT small number parameter to user
This commit exposes a real number to the user that affects FCT. Example:

fcteps 1.0e-7

The above will use that number below which partial FCT quantities are
treated as zero, instead of machine zero. Besides sysfctvar, which
allows skipping a field when computing the system limit coefficient for
a cell, fcteps provides another way to ignore small but noisy fields.

Increasing fcteps can also be used to achieve reasonably matching
results in parallel in cases where the machine zero makes this very
hard. This should be used with caution though, because increasing fcteps
to a too large value can great screw with the solution quality.

The default is std::numeric_limits< tk::real >::epsilon(), which on my
workstation happens to be 2.22045e-16.