Compute number of points across whole problem
This commit puts in communication to compute the correct number of
points in parallel both before and after initial mesh refinement.

The communication uses Charm++ custom reducers for reducing a hash set
(std::unordered_set) and counts the number of globally unique node ids
across the whole partitioned problem. The reduction relies on Charm++'s
tree-based communication and uses DAGs to detect whether both the number
of elements as well as the number of nodes have been computed.