Use Discretization as a shadow when initiating a reduction for node diagnostics
Instead of Refiner. Experimented with destroying Refiner if dtref is
false to free memory but that did not work out. However, keeping this
change because it is simpler to use Discretization instead of
Discretization->Refiner, and we may be able to destroy Refiner in the
future. IT does not seem to cause any problems.