Replace all-to-all in Refiner with a quyer-response
Thsi commit replaces an all-to-all communication in Refiner, used to
compute shared edges between new refinement steps, with a two-step
query-response algorithm, similar to how Sorter computes node
communication maps after mesh partitioning. This scales better, faster,
and as it turned out, we don't even need the shared edges, we only
needed which chare a give edge is shared with. Thus this not only
eliminates snowballing all boundary nodes targeting all chares, but also
reduces the data that needs to be communicated, leading to huge
improvement in parallel and for large meshes.