ALECG discretization appears to be correct
However, we are still not quite there yet: (1) the nodal gradients only
implemented in serial, and (2) there appears to be still a problem at
the boundary when running the vortical flow problem. I have a good idea
how to do (1), but (2) needs to be hunted down first.

This commit ended up huge. Here is current state and details:

Temporal (RK3) and spatial discretization for the Euler equation appear
to be correct. This is measured by a couple of identities being
satisfied.  These are Eq. (A.21) and D_i^{vw} = - D_i^{wv} in Waltz, et
al, "A three-dimensional finite element arbitrary Lagrangian-Eulerian
method for shock hydrodynamics on unstructured grids", Computers &
Fluids 2014.

Added calculation of point normals, which should also work in parallel.
This will be needed for setting symmetry (free-slip) velocity BCs.
Merged in branch 'norm', which has the same for DiagCG, which is now
also in ALECG.

Changed how the elements surrounding edges data structure is presented.
Previously, esued simply stored the node ids that made up all edges.
However, the algorithm we are now implementing requires an outer loop
over edges and an inner loop over elements surrounding edges. While we
did have data structures for for such a loop, they were not intended to
be used this way: the edge-end point indices were not ordered in a way
that could be used to access the element node ids of elements
surrounding edges as well as the edge end-points of the edge. Now we
assign a list of surrounding element ids to a node-id pair (an edge) in
esued. This also required changes to some of the unit tests and also
eliminated some unit tests.

Moved the triangle area calculation in DerivedData.[ch]pp to its own
routine so we can call it on the fly for boundary triangles (only). So
far this was only possible to use it together with other data, e.g.,
cell centroid and face normals, used for DG.