Remap ALECG::m_triinpoel to local ids
Also, change what it contains: instead of the nodes on the partition
boundary, now it contains nodes on the physical boundary on which the
user has set boundary conditions. This has two implications:

1. The user must absolutely now set BCs on all of the problem boundary
(as a union of possible multiple side sets yielding a non-leaky
boundary), and

2. ALECG::m_triinpoel now contains (local) node ids of only the physical
boundary nodes. This is what is correct in parallel, because this data
structure is used to compute the boundary integral terms and those are
nonzero only along physical boundaries and zero along partition
boundaries.

While this works in parallel with transport/advection, the solution
computed is not yet correct.  Even though the nodal gradients as well as
the fluxes are now communicated (hopefully correctly) in parallel, I
think one thing that's still missing is that computing the
dual-mesh-face normals also requires communication, which is not yet
done.