Commincate normals in ALECG
This does not change the serial result. In parallel, however, there
appears to be a problem along chare-boundary edges that still needs to
be debugged.

Setting up chare-boundary edges uses an all-to-all, snowballing all
boundary edges to all chares. This will not scale to large problems and
large core counts and thus will have to replaced with a more efficient
point-to-point communication. (This comes from Refiner, which will also
have to be replaced.) For now, however, determining the chare-boundary
edges appears to be correct, because all edges that are sent are also
found by all receivers.

Note that relying on the point communication map
(Discretization::m_msum) by finding both edge-end points going to
another chare does now work because it misses some edges, probably some
corner cases.