Partitioner now generates edge comm map
This is a relatively large change-set but at a high level, it is really
not a complicated one, only a lot of code is touched.

Background:

In an attempt to generate an edge-communication map for ALECG in a way
that scales, this commit piggy-backs the point-to-point query/response
algorithm in Sorter to generate not only the node communication map but
an edge communication map as well. I am putting the edge communication
calculation in Sorter because that already has a query/response
algorithm setup that we know scales well. Another option would have been
to do this in ALECG. That would also have made sense since the edge comm
map will only be needed by ALECG (for now). However, that solution would
have required two reductions to Transporter, which then would have
required empty stubs for all other child schemes, DiagCG, DG.
Furthermore the query/response algorithm would have almost exactly
looked like the one in Sorter anyway. The price for the choice of doing
this in Sorter is that now the edge comm map is stored in
Discretization, which is not really needed for DiagCG and DG. Perhaps we
can turn off computing and storing the edge comm map for those two in
the future. For now, I just kept it in there.

Another minor change is that we now only store the node communication
maps as a unique set of a nodes associated to chare ids, instead of a
vector. Both DG and Refiner required the node comm map as a map of sets,
while DiagCG, DistFCT, and ALECG needed this as vectors. So far
Discretization stored this as vectors. Now Discretization only stores is
as a map of sets and the sets are converted to vectors on the fly as
needed. This probably adds some performance penalty during time
stepping. Will have to benchmark this in the future to see how much of a
hit this means, but I doubt it is significant. In exchange the size of
the states of DG and Refiner is reduced.

There is also a few renames: there is almost no "msum" left anymore.
That was not a very good name anyway. It stood for "mesh surrounding
mesh", which really stored a symmetric node communication map. Now the
node communication map is called a nodeCommMap and msum is reserved for
a map that associates ALL maps (node and edge) to chare IDs, while this
is necessary, e.g., in Sorter. In Discretization these comm maps are
stored separately to easily interface with existing code: m_nodeCommMap
and m_edgeCommMap. (This latter is not yet used/tested.)

I also added a new file, src/Mesh/CommMap.hpp, which contains some type
aliases for these maps so the underlying types are easier to change in
the future and most importantly, to make it easier to add new comm maps
in the future.

Next is to see if the new edge comm map works in ALECG. (And of course
to remove the existing all-to-all in ALECG for the computing the edge
comm map there.)