Clean gradients receive buffer in ALECG
This makes ALECG::Transport work in parallel.

There are other minor cosmetics/whitespace changes in this commit, as
well as some code in Sorter that ensures that the edge comm map keeps
only those edges in whose both end-points are in the node comm map. This
latter did not make a difference for the case I have been testing but
this appears "more correct".