Store nodal gradients only in chare-boundary nodes
This commit reduces the side of ALECG::m_grad from all points to only
the points on the chare boundary. This will help load balancing when a
smaller object state must be migrated.

The gradients are now computed in two stages: (1) the chare-boundary
nodes and (2) internal nodes. Naturally, only the chare-boundary nodes
need to be communicated, which is done first. Then when the rhs is
computed and the gradients are needed in all nodes, storage is allocated
temporarily, the chare-boundary gradients are copied in, followed by
computing the internal node gradients. After using the gradients the
data is thrown away.

The performance hit was tested with two implementations with a 230K-cell
mesh on 6 cores, comparing to the version with storing the gradients in
all mesh points:

(1) cg::Transport::egrad() inline, repeating code, which yields about a
2% improvement(!) in performance, and

(2) calling cg::Transport::egrad(), which reuses code at three places,
which yields a 2% degradation in performance.

Either of the above is acceptable considering the gains in reduced
memory usage (and its benefits to migration and load balancing),
especially problems with a lot more than a single scalar equation.