Convert sysfct from vector< bool > to vector< int >
First, thanks to @adityakpandare for tracking down this problem. Since
the input to sysfct from the user has so far been stored in a vector<
bool > and apparently Charm++'s serialization corrupts this data in
parallel (perhaps because std::vector< bool > is not really a real
container, not every PE has seen the correct user setting for this input
and thus the results got corrupted.

This commit fixes this by chaning the type of the container in inciter's
input deck for sysfct to std::vector< int > and adding a function in the
grammar that still allows parsing for true or false but storing the data
as integers.

This, as expected, changes the regression test where this is enabled and
has an important effect. Also tested a larger mesh using sysfct in
parallel and the improvement is huge: now the serial and parallel
results are pretty close, however, not as closes as being able to use
the same diagnostics file for a regression test exercising
virtualization in parallel.