Fix ALECG::m_bnorm
A few commits ago, I made the mistake of keeping the keys of m_bnode as
global ids, but that is wrong because when we search for nodes from
triinpoel we need local ids and not global ids, since I forgot that
ALECG ctor remaps to triinpoel to local ids.

Keeping m_bnorm using local ids helps keeping this consistent across
ALECG and DiagCG.

This commit also applies symmetry BCs on the initial conditions, which
was not done previously.

Sod shocktube appears to work correctly in parallel now, setting sym BCs
correctly.