Put back esued-based flux computation to ALECG
This commit puts back the esued-based, (esued = elements surrounding
edges) flux computation in ALECG Transport and CompFlow. While the
results looked suspiciously alright using the previous version, neither
Transport nor CompFlow could reproduce the same L2 norms in serial vs.
parallel. I suspect that this is because the computation of the
dual-face normals and the flux contributions do not commute, but assumed
in the previous version.

This commit puts back the flux computations in the domain-edge terms,
spelling out the Rusanov flux and computing the wave speed, that closely
follows Jacob's paper, in both Transport and CompFlow.

Advection of the Gauss hump and the Sod problem appear correct and now
compute the exact same result in serial, parallel, and parallel +
virtualization.

Though currently unused, kept src/PDE/Integrate/Riemann/Rusanov.hpp,
since it is written in a way that could be reused in DG.