Store ALECG dual-face normals only along chare boundaries
Instead of storing dual-face normals associated to all edges, ALECG now
stores the dual-face normals only in chare-boundary edges (which need
communication to be complete). The chare-boundary data is zero in serial
and significantly smaller in parallel compared to storing them all as
before. When the normals are needed, we now recompute them.