Expose screen output file name configuration to all command line parsers
This commit enables configuring the name of the screen output log file
in all executables. IThe default is "screen.log" and the command line
argument is -g or --screen. (We are somewhat running out of unique
single-letter command-line arguments.)

It also adds the option to pass in an file open mode into tk::Print so
that appending to the screen output log file is possible. This is useful
because sometimes we re-create the printer object as needed, either
because we don't always want to carry it around or just don't want to
(or cannot) store it (e.g.,because migrating references is problematic).
In those cases appending to the screen output is useful so logging can
continue in the same file.

tk::Print now also no longer touches its verbose stream object
(m_stream), setting up a tee buffer, unless a filename is configured.
The assumption is that if the screen output log file name is empty in
tk::Print's constructor, the client does not care about logging the
stream to file, so we don't even bother setting the tee buffer up. (It
would also fail some regression tests that require saving and restoring
the verbose stream state.)