Add simple TaggedTuple printer
This can do simple (unformatted) output by iterating through a
TaggedTuple using brigand::for_each. Simple, because everything goes
onto a single line. Two versions a defined: an operator<<() and a
print() function. The first one simply prints the whole tuple, while the
second one allows for a list of types (tags) whose data will not be
printed. This latter will be useful to selectively ignore some data when
printing CmdLine or InputDeck objects.

Note that those tags in the TaggedTuple whose associated data is to be
printed, must define a static function name(), returning a string that
is characteristic of the name of the tag, e.g., instead of

struct output {};

now we need

struct output { static std::string name() { return "output"; } };