Replace constexpr with SFINAE for enum class operator<<
This works better, because the previous version can be compiled in with
all non-enum-class types, executing the else-branch of the constexpr,
which, surprisingly, can compile without an error for arbitrary types,
such as a std::vector, even if there is no operator<< defined for it.
This is bad, since that simply results in a segfault at runtime.

The SFINAE version works better, because if the type is not an enum
class, the operator<< is removed from the overload resolution set at
compile time, and thus correctly generates a compile-time error for
types that do not define operator<<.