Set EXCEPTIONS_WRITE_TO_CERR based on CHARM_SMP
Writing to std::cerr from tk::Exception is not thread-safe in Charm++'s
SMP mode. This commit ensures the correct setting, instead of relying on
the user to set it (or forget it) during a cmake call. Note that
overriding is still possible by passing -DEXCEPTIONS_WRITE_TO_CERR=false
or -DEXCEPTIONS_WRITE_TO_CERR=true to cmake.

This eliminates the need to set it on the cmake command line, so we
remove its setting on Azure.