Do not recompute dual-face normals every time step stage in ALECG
This commit changes the strategy of how we store and compute the
dual-face normals in ALECG. Previously, only the normals associated to
chare-boundary edges were stored (as those requires communication) and
the rest were recomputed (at every time step stage). If the mesh does
not change, recomputing the normals is unnecessary.

With this commit we now store the dual-face normals for all edges, and
communicate them as needed along chare-boundaries as before, but no
longer recompute them at every time step stage. This gains about 20%
speed-up on a 750K mesh on 30 CPUs.  However, this stores a larger data
in ALECG, so a price is paid if/when migration is required for load
balancing. However, the normals do not necessarily have to be sent
across the network with migration but can be recomputed. This last step
is not yet done, but will most likely provide the best of both worlds.
However, this will need to be compared to sending the data first.

This commit also moves the location at which the dual-face normals and
the boundary-point normals are computed, from the constructor to (later)
when the lhs is computed. This makes for a better organization, since
if/when the mesh changes, e.g., after AMR, the lhs needs to be
recomputed and so now we also recompute the two types of normals at the
same time (overlapped and independently from the lhs). Though ALECG is
not yet tested with AMR, this step will be required for that.