Move dt out of DiagCG::rhs()
This commit moves the dt multiplier outside of the terms on the rhs for
DiagCG. This affects both cg::Transport and cg::CompFlow. The terms of
the rhs are broken up into gather and scatter and also to terms
independent of dt and those that are not. This will allow scheduling
tasks differently depending on the availability of dt.