Group cells in DiagCG by boundary and internal
This commit groups the elements on which the right hand side operators
are computed in DiagCG by those that contribbute to nodes on the
chare-boundary and those that do not, i.e., internal ones. This allows
creating tasks for working on these two groups of cells separately.

We compute the boundary ones first (which is quick), send them out for
communication, then proceed to the itnernal ones. This yields a much
better CPU utilization, as tasks are finer grained. Tested at 900 and
14,400 CPUs, this does not improve performance, measured by wall-clock,
likely partly because the current algorithm visits the chare-boundary
cells twice and thus must recompute gradients, etc.

Next is to improve on this by not having recompute boundary cell
contributions twice.