Partition elements in DiagCG to bnd and int
This commit partitions the element list in DiagCG into those that
contribute to chare-boundary nodes (bnd) and to those that contribute to
internal nodes only (int). This allows improving the existing algorithm
by a factor 15%, since the chare-boundary element contributions are no
longer computed twice (now a cell is only visited once), and also
removes the if-test at a very low level of the oprators that decides
whether to contribute to a node or not, and also simplifies some
function interaces and additional code.

Also decrease the LB frequency for a regression test to make it run
faster.