Add optional operator-access mesh node reordering to DiagCG
This commit adds, and exposes to the user, an optional local mesh node
reordering that is based on the PDE operator access pattern to DiagCG. A
similar idea can also be implemented into other schemes.

In the input file, we also rename "reorder" to "pelocal_reorder",
since this new can now be configured by "operator_reorder".

The default for both types of reordering remains false.