Implement resize and push_back for Data's EqCompUnk access
This are only stubs, that define implementations for these two member
functions for tk::Data. The rationale is that at least the code now
builds for both field-, as well as, equation-major access data access.
However, for the latter, these two operations would be inefficient, so
they remain unimplemented. These are only called after AMR, which not
always exercised, and now throw if someone tries to call these
functions. Nevertheless, this enables testing the equation-major access
patter (without AMR), which is still useful.