Revert both gather/scatter and bnd/int splits in DiagCG
This commit reverts both splits in the DiagCG algorithm: (1) The
gather/scatter split was separating work for the two steps (gather and
scatter) of a time step, and (2) work on the boundary and internal
elements.

The idea was that with gather not dependent on the availability of the
new dt, the gather step can be speculatively started earlier, before the
new time step, allowing more overlap and thus better CPU utilization.
While this is realized, the split requires some extra data structures,
their setup, and the actual separating of the gather and the scatter
steps, which incurs overhead.

The other split that is reverted is the work on boundary cells, followed
by communication, then work on internal cells. The idea here is that the
there are less boundary cells, which if done quick, communication is
sent out earlier compared to an unsplit algorithm and this allows more
overlap. This is also realized.

However, the both of the above splits together incur just enough
overhead, more complex, code and thus perhaps cache misses (due to the
split) that grind-time slows down by about 3-4% on a single node (using
a 18M-cell mesh) and by about 13-15% on 10 compute nodes (using a
800M-cell mesh). These results are fairly reproducible, also tested from
360 CPUs to 30K CPUs and consistent and so this commit reverts these
both splits. As a result, the code is somewhat simpler and faster,
allows less overlap, and shows worse CPU utilization.