Add stagnation point BC to DiagCG
This adds stagnation-point BCs to inciter's compflow for DiagCG. This is
now possible:

  compflow
    ...
    bc_stag
      point 0 -2.667 -23.2918 end
      radius 1.0e-8 end
    end
    ...
  end

specifying a point in 3D space and a radius around it. Note that
multiple points can be specified with multiple/independent radii. If a
mesh node falls within the radius around the point, the velocity of the
point will be fixed initially and throughout time stepping. This is now
hooked up for DiagCG only.

Also add a regression test that exercises setting a stagnation point.