Detect if just returned from a checkpoint and if so, zero timers
This is somewhat tricky, so here is some more background:

The only way I have found detecting in Charm++ the fact that the
executable has just been restarted from a checkpoint is Tom Quinn's
solution, explained at
https://lists.cs.illinois.edu/lists/arc/charm/2020-03/msg00005.html,
which is exactly what we follow in this commit.

The indicator variable, used to detect the fact that the executable has
just been restarted, is stored in

inciter::g_inputdeck.get< tag::cmd, tag::io, tag::nrestart >()

i.e., the numer of restarts and is incremented in the Main chare
migration constructor, called after returning from a checkpoint.

This writes to the global state in g_inputdeck and is only available on
PE 0. (We assume that the Main chare always resides on PE 0.) This is
(hopefully) further discussed followed by
https://lists.cs.illinois.edu/lists/arc/charm/2020-03/msg00007.html.

The we pick up nrestart from g_inputdeck in Transporter::resume().
Transporter always resides on PE 0, see InciterDriver::execute().

Transporter::resume() then broadcasts nrestart to all Discretization
chares and thus makes it available to all child schemes. Discretization
stores nrestart in m_nrestart and does not mess with g_inputdeck, so
nrestart remains inconsistent in parallel on PEs other than PE 0.

<Scheme>::evalLB() then calls Discretization::restarted(), which figures
out, based on (a potentially new incoming) nrestart and its existing
m_nrestart if there was a restart from a checkpoint or evalLB was just
called after doing a checkpoint (without a restart). If
Discretization::restarted() arrives at the result that it has been
restarted, it updates its copy of nrestart (m_nrestart) and zeroes the
timer measuring the time stepping wall-clock time and the grind time,
which was the whole reason these changes have been implemented. As a
result, the ETE, ETA, and EGT on-screen columns are correct after a
restart. Previously, these results were based on time-stamps saved
during a previous run, so they contained the times elapsed between
restarts, which is wrong and pretty much useless.

This commit also puts in zeroing the grind time for DG, which has not
been in place yet.

DG is not yet tested with checkpoint/restart.

Now it might even be possible to automatically name the log file name
based on the number of restarts, without having to look at the disk,
e.g., listing the number of times the log file exists, which is pretty
error-prone. The log file name can be configured if/when #394 is merged,
but automatic naming would be more user-friendly.