Compute new log file name automatically for each restart
This affects all executables, because part of this is generic code
reused among all executables, though only inciter can do restarts for
now.

It is still possible to set the log file name via -O <filename>, in
which case we don't mess with it. But if this is not done, now we
determine how many times the run was restarted from a checkpoint and
postfixes the default log file name with that number.

For example: one runs inciter once, which also creates a checkpoint:

$ inciter -i ... -o ...

This creates 'inciter_screen.log' by default, and the checkpoint is
created in directory 'restart/'.

Then doing a restart with that checkpoint:

$ inciter -i ... -o ... +restart restart

will continue, but now will not overwrite the log file
'inciter_screen.log', but instead will write to 'inciter_screen.log.1'.

The next restart:

$ inciter -i ... -o ... +restart restart

will write to 'inciter_screen.log.2', and so on.