Switch the order of filling the node gradients
This commit switches the order in which the gradient data structure is
populated in ALECG. This is in function nodegrad in src/PDE/CGPDE.cpp.

This allows eliminating a search in the loop over internal points, since
the chare-boundary points will simply overwrite those. This results in a
small performance increase and a reduction of some code complexity,
since the global ids now do not have to be passed down to that function.