Add buildtype-based exception spec for [c]ref_find()
This commit adds conditional exception specification to the ref_find and
cref_find functions. The previous behavior was that even in DEBUG, these
functions threw an exception if the key was not found in the map
searched. Now no exception will be thrown in RELEASE if the key is not
found, only in DEBUG. This migh speed up some loops, but most
importantly, this is more consistent with th expected behavior of no
exceptions in RELEASE and throwing exceptions only in DEBUG.