Vectorize ALECG's boundary integral loop
This commit rewrites the boundary integral in ALECG's CompFlow and
vectorizes it. Besides this it also has a few improvements:

* ALECG::m_triinpoel now stores local node ids. This eliminates some
global->local id lookups.

* It removes unnecessary tk:: namespace specifications from
DerivedData.[ch]pp.

* It adds new vectorized overloads to the area() and normal() functions
in DerivedData.[ch]pp.

* It simplifies searching for node ids at which symmetry BCs are set in
ALECG. The new data structure, ALECG::m_symbcnodes, is a vector which
has 1 if the node has a symmetry BC, and 0 if not. This eliminates
searching in a map, and thus possible to use in a vectorized loop. It
also does not carry the boundary normals for this purpose, which was
unnecessary, previously done using AELCG::m_bnorm.

* Provides vectorized versions of eos_pressure().

Some regression tests will need adjustment, but this is deferred to when
all important loops have been vectorized, so only a single update of the
baselines is necessary.