Revamp src
This commit does a partial revamp of how we use optional source terms.
This is basically a WIP towards a vectorizable source integral in ALECG,
but since the source terms a connected across the schemes and problem
types, it affects all interfaces. This is an attempt to make it a little
more specific to the type of the PDE. ALECG's source integral is not yet
vectorized, because the intel compiler segfaults on calling a static
member function from a vectorized loop. So that part is left for future
work, nevertheless the changes here are a good idea IMO.