Fix 2 stagnation point bugs
The first bug fixed was in DiagCG as stagnation points have been
applied. It was writing values to an array way beyond the allocation.

The send, more subtle, bug fixed was that CGCompFlow was storing a
newly introduced state m_stag, which was written to initialize(), after
the constructor. This is not allowed, because these equation objects are
created very early on and made global adn thus read only. This led to
problems in SMP mode only. This was fixed by a slightly more expensive
way of testing for stagnation points, based on user config parsed. This,
however, does not affect performance in a measurable way, and every loop
that calls the new CompFlow::stagPoint() function is still vectorized
correctly.

With this all regression tests, except for ALECG/CompFlow, pass again.