Delay applying symmetry BCs in ALECG
This fixes a bug in applying symmetry BCs in ALECG on the initial
conditions. Previously, the symmetry BCs were applied immediately after
the ICs were set. However, the boundary point normals are not yet
available at that point and so no symmetry BCs were applied. This commit
delays the application of the symmetric BCs to after the boundary-point
normals are available so that the symmetry BCs are indeed applied on the
ICs. This also required moving the initial field output after this,
hence the slight reorganization of the logic within ALECG

Somewhat surprisingly, this change does not need updates on the
baselines that set symmetry BCs. (Only those ALECG tests fail that apply
operator reordering, which is currently buggy.)