Add local time stepping to ALECG
By default this is off, and can be enabled by, e.g.,

inciter
  ...
  steady_state true
  residual 1.0e-8
  ...
end

Note that if marching to steady state is on, 'nstep' and 'term' as still
considered when deciding on the end of time stepping: either of nstep,
term, or the L2-norm of the residual reaching thei respective condition,
time stepping will stop.

No regression tests, but this is so far tested with vortical flow and
the Onera M6 wing, for both this makes a difference in convergence rate
when a final steady state can be assumed. Note that we do not deactive
any nodes, but local time stepping allows signals to propagate faster
based on local time step sizes.