Move query function for bc nodes from PDE to Discretization
This commit moves and generalizes the function that is used to query the
node ids at which a particular boundary condition type is configured by
the user. This has been in cg::[CompFlow,Transport]::symbcnodes(), and
now it is moved to Discretization::bcnodes().

Discretization::bcnodes() is now also a template, and does a
compile-time loop to query all PDE types available. This way this can
remain generic, and can be called directly from DiagCG and ALECG. Also,
since the code can be parameterized by the BCType, it can be used for
finding nodes for symmetry or farfield BCs, since both need those list
of points, e.g., to find boundary-point normals.