Housekeeping before hooking up farfield BCs into ALECG
* Abstract away the computation of boundary-point normals so the same
code can be reused between DiagCG and ALECG. There is now a new function
in PDE/CGPDE.[ch]pp: bnorm(), which does this.

* ALECG now maintains separate unique sets of node ids for symmetry BCs
and farfield BCs: these are m_symbcnodes and m_farfieldbcnodes. Boundary
point normals are now computed on the union of the symmetry and farfield
BC node lists in ALECG.

* Setting symmetry BCs in cg::CompFlow::symbc() is now based on the
unique set of nodes at which symmetry BCs are set and not for which
boundary-point normals are computed (which now can be larger).

* cg::Transport now does not need symbc(), due to f3d0a1daf.