Improve on farfield symmetry BCs
This commit improves on how symmetry and farfield BCs are applied in
ALECG and DiagCG.

Boundary-point normals are now computed separately for sidesets at which
BCs are applied. This allows the correct and different normals even if
multiple BCs are assigned to a single point as the point could belong to
multiple side sets, for example for symmetry and farfield BCs. This has
implications on the data structures for boundary-point normals, on how
BCs are applied, and also affects DiagCG.

This also required some updates to the baselines, since now, for
example, the symmetry BCs are "more correct" (TM) for the Sod problem on
a rectangular cross-section mesh, where previously normals along edges
were 45 degrees and now they are applied multiple times on top of each
other for different side sets using the correctly oriented normals.

This commit appears to correctly compute the Onera wing using local time
stepping with ALECG.