Apply symmetry and farfield BCs on IC for diagnostics
This commit applies the symmetry and farfield BCs on the initial
conditions, which are evaluated in place of the analytic solution if
that does not exist for the problem type. This is evaluated for
computing diagnostics, e.g., L2-norms of the difference between the
analytic solution (or IC) and the current solution.