Fix boundary point normal calculation
This commit fixes a bug in the boundary point normal calculation,
affecting both DiagCG and ALECG symmetry and farfield BCs.

Previously there was cross-contamination between side sets when
computing the normals using the normals of boundary face triangles
surrounding a boundary point. The result was that the normals were wrong
associated to a boundary point that belonged to multiple side sets,
independent of the type of boundary condition specified on it.

Updated Sod regression test baselines for nodal schemes.