When receiving boundary-point normals do not assume own contribution
Not sure if this has so far made any difference, but with this change
the boundary-point normal communication appears "more correct" (TM).
Basically, this change ensures that there is no need for an own
contribution to a boundary point when we combine the contributions to a
boundary point normal. I imagine that this may happen if a chare only
has a single boundary point (or two) on a side set that requires
normals, but no own contributions exist, since that is always based on a
side set face (at least a single triangle face) on a chare. This commit
ensures that contributions to such boundary-point normals are not
ignored on the receive side.

Update the single regression test baseline where this apparently
mattered, causing a minor diff in the field output only (diagnostics and
history output did not change).