Add ids to history points
Previously, we configured history points as

   history
     point 0.1 0.05 0.025 end
     point 0.9 0.05 0.025 end
   end

Now we require a point identifier as

   history
     point p1 0.1 0.05 0.025 end
     point p2 0.9 0.05 0.025 end
   end

where the token after 'point' keyword must be a valid C identifier.

We use the point ids to name the history output files. Previously, we
used the coordinates as

  out.hist.{0.9_0.05_0.025}
  out.hist.{0.1_0.05_0.025}

Now this is as

  out.hist.p1
  out.hist.p2

which is more user-friendly.