Add partial lua parsing for inciter::compflow::ic
This commit adds code to interpret embedded Lua code inside inciter's
compflow block. This is experimental at best, but appears to work,
partially interpreting the ic ... end block as lua code.

As an example, instead of

inciter
  ...
  compflow
    ...
    ic
      density 1.0 end
      velocity 0.0 0.0 0.0 end
      pressure 1.0 end
      box
        xmin  0.5 xmax 1.5
        ymin -0.5 ymax 0.5
        zmin -0.5 zmax 0.5
        density 0.125 end
        pressure 0.1 end
      end
    end
    ...
  end
  ...
end

now this is possible:

inciter
  ...
  compflow
    ...
    lua
      ic = { density = { 1.0 },
             velocity = { 0.0, 0.0, 0.0 },
             pressure = { 1.0 },
             box = { xmin =  0.5, xmax = 1.5,
                     ymin = -0.5, ymax = 0.5,
                     zmin = -0.5, zmax = 0.5,
                     density = { 0.125 },
                     pressure = { 0.1 } } }
    end
    ...
  end
  ...
end

(The above is a typical example of configuring Sod's problem using
background IC and an overlaid box specifying values different from the
background.)

The above two are equivalent. Lua parsing happens after the built-in
PEGTL parsing, so lua blocks can overwrite PEGTL-parsed blocks,
independent of where the data is in the input file. No significant
error checking is done for lua code.

Also added a regression test, exercising lua input. As a result, from
this commit lua and sol2 are now required as third-party libraries.